Bugzilla 4601: Change ElemNumber to use a long integer for the value of the number. This will allow us to format bigger numbers. As far as japanese multipliers, we decided to document as a xalan limitation the fact that the last 2 japanese multipliers will not be supported. If anyone has a strong and valid user case for this, we will reconsider. For now, we could not justify the expense of moving to BigInteger to support these multipliers.
git-svn-id: https://svn.apache.org/repos/asf/xalan/java/trunk@335800 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/xml/utils/res/XResources_zh_TW.java
@@ -111,7 +111,7 @@ public class XResources_zh_TW extends XResourceBundle
     { "zero", new char[]{ 0x96f6 } },
 
     //These only used for mutiplicative-additive numbering 
-    { "multiplier", new int[]{ 100000000, 10000, 1000, 100, 10 } },
+    { "multiplier", new long[]{ 100000000, 10000, 1000, 100, 10 } },
     { "multiplierChar",
       new char[]{ 0x5104, 0x842c, 0x4edf, 0x4f70, 0x62fe } },
     { "digits",
